refactor: useDragDrop step 6 — desktop drop on calendar and date header
